add current details fragment

+++ app/src/main/kotlin/ru/stolexiy/openweather/data/remote/model/current/RainDto.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class RainDto(
     @SerializedName("1h")
-    var volume1h: Int? = null,
+    var volume1h: Double? = null,
     @SerializedName("3h")
-    var volume3h: Int? = null
+    var volume3h: Double? = null
 )
